removed unnecessary import and moved mapping of lists for dropdowns into container. Also fixed ports itemType
