+++ src/Persistence.LinnApps/Repositories/ImportBookRepository.cs
@@ -3,7 +3,6 @@
     using System;
     using System.Linq;
     using System.Linq.Expressions;
-    using System.Security.Cryptography.X509Certificates;
 
     using Linn.Common.Persistence;
     using Linn.Stores.Domain.LinnApps.ImportBooks;
@@ -21,11 +20,8 @@
 
         public ImportBook FindById(int key)
         {
-            return this.serviceDbContext.ImportBooks.Where(b => b.Id == key)
-                .Include(x => x.InvoiceDetails)
-                .Include(y => y.OrderDetails)
-                .Include(z => z.PostEntries)
-                .ToList().FirstOrDefault();
+            return this.serviceDbContext.ImportBooks.Where(b => b.Id == key).Include(x => x.InvoiceDetails)
+                .Include(y => y.OrderDetails).Include(z => z.PostEntries).ToList().FirstOrDefault();
         }
 
         public IQueryable<ImportBook> FindAll()
@@ -50,8 +46,7 @@
 
         public IQueryable<ImportBook> FilterBy(Expression<Func<ImportBook, bool>> expression)
         {
-            return this.serviceDbContext.ImportBooks.Where(expression)
-                .Include(y => y.OrderDetails)
+            return this.serviceDbContext.ImportBooks.Where(expression).Include(y => y.OrderDetails)
                 .Include(x => x.InvoiceDetails);
         }
     }
